requirements.txt
pillow
argparse
bilibili-api-python